+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Fix Accessibility problems in ProfileLinker Component (settings/data-import/job-profiles) (UIDATIMP-434)
 * Fix an error occurred while searching for associated profiles (UIDATIMP-769)
 * Status descending sort on Data Import home page not working. Fixed (UIDATIMP-758)
+* Fix an error on switching existing record types on Match profile Create/edit screen (UIDATIMP-804)
 
 ## [3.0.3](https://github.com/folio-org/ui-data-import/tree/v3.0.3) (2020-11-13)